requirements.txt
notebook
pandas
scikit-learn
psycopg2-binary
apache-airflow
apache-airflow-providers-amazon
mlflow